refactor(runtime): split mcp_stdio into server_manager and process modules
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
