+++ crates/runtime/src/mcp/mod.rs
@@ -1,9 +1,11 @@
 pub mod client;
 pub mod naming;
+pub mod process;
 pub mod server_manager;
 pub mod types;
 
 pub use client::*;
 pub use naming::*;
+pub use process::*;
 pub use server_manager::*;
 pub use types::*;
